+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.1-alpha.1",
-  "contentSha256_12": "eafa205c505b",
+  "pkgVersion": "1.1.1-alpha.2",
+  "contentSha256_12": "86a32e832945",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.1-alpha.1.js"
+  "hacsFile": "nodalia-cards-1.1.1-alpha.2.js"
 };
-export const pkgVersion = "1.1.1-alpha.1";
-export const contentSha256_12 = "eafa205c505b";
+export const pkgVersion = "1.1.1-alpha.2";
+export const contentSha256_12 = "86a32e832945";
 export const file = "nodalia-cards.bundle.js";